Rename extent.h -> edata.h.
This name is slightly pithier; a full-on rename will come shortly.
