+++ src/edata.c
